Refactor concat / combine / merge into `xarray/structure` (#10134)
* Refactor concat / combine / merge into `xarray/structure`
