+++ xarray/__init__.py
@@ -28,10 +28,7 @@ from xarray.computation.computation import (
     where,
 )
 from xarray.conventions import SerializationWarning, decode_cf
-from xarray.core.alignment import align, broadcast
-from xarray.core.combine import combine_by_coords, combine_nested
 from xarray.core.common import ALL_DIMS, full_like, ones_like, zeros_like
-from xarray.core.concat import concat
 from xarray.core.coordinates import Coordinates
 from xarray.core.dataarray import DataArray
 from xarray.core.dataset import Dataset
@@ -44,7 +41,6 @@ from xarray.core.extensions import (
 )
 from xarray.core.indexes import Index
 from xarray.core.indexing import IndexSelResult
-from xarray.core.merge import Context, MergeError, merge
 from xarray.core.options import get_options, set_options
 from xarray.core.parallel import map_blocks
 from xarray.core.treenode import (
@@ -55,6 +51,10 @@ from xarray.core.treenode import (
 )
 from xarray.core.variable import IndexVariable, Variable, as_variable
 from xarray.namedarray.core import NamedArray
+from xarray.structure.alignment import align, broadcast
+from xarray.structure.combine import combine_by_coords, combine_nested
+from xarray.structure.concat import concat
+from xarray.structure.merge import Context, MergeError, merge
 from xarray.util.print_versions import show_versions
 
 try:
