+++ xarray/computation/weighted.py
@@ -9,9 +9,9 @@ from numpy.typing import ArrayLike
 from xarray.computation.apply_ufunc import apply_ufunc
 from xarray.computation.computation import dot
 from xarray.core import duck_array_ops, utils
-from xarray.core.alignment import align, broadcast
 from xarray.core.types import Dims, T_DataArray, T_Xarray
 from xarray.namedarray.utils import is_duck_dask_array
+from xarray.structure.alignment import align, broadcast
 
 # Weighted quantile methods are a subset of the numpy supported quantile methods.
 QUANTILE_METHODS = Literal[
