+++ xarray/plot/dataset_plot.py
@@ -6,7 +6,6 @@ import warnings
 from collections.abc import Callable, Hashable, Iterable
 from typing import TYPE_CHECKING, Any, TypeVar, overload
 
-from xarray.core.alignment import broadcast
 from xarray.plot import dataarray_plot
 from xarray.plot.facetgrid import _easy_facetgrid
 from xarray.plot.utils import (
@@ -16,6 +15,7 @@ from xarray.plot.utils import (
     _process_cmap_cbar_kwargs,
     get_axis,
 )
+from xarray.structure.alignment import broadcast
 
 if TYPE_CHECKING:
     from matplotlib.axes import Axes
