+++ xarray/structure/alignment.py
